fix: fixed up code
