+++ projects/01_fyyur/starter_code/templates/forms/edit_venue.html
@@ -30,7 +30,7 @@
       <div class="form-group">
         <label for="genres">Genres</label>
         <small>Ctrl+Click to select multiple</small>
-        {{ form.genres(class_ = 'form-control', placeholder='Genres, separated by commas', id=form.state, autofocus = true, value = venue.genres) }}
+        {{ form.genres(class_ = 'form-control', placeholder='Genres, separated by commas', id=form.state, autofocus = true) }}
       </div>
       <div class="form-group">
           <label for="genres">Facebook Link</label>
